Reuse form templates and structures
